fix(playerbot): Resolve 1,197 compiler warnings for clean build
Comprehensive warning cleanup reducing total warnings from 1,212 to 15:

Warning fixes by category:
- C4189: Fix 78 unused initialized variables (removed or utilized)
- C4101: Fix 9 unreferenced local variables
- C4305: Fix 38 double-to-float truncations (static_cast<float>(M_PI))
- C4018/C4389: Fix 29 signed/unsigned comparison mismatches
- C5038: Fix 10 member initialization order issues
- C4099: Fix 20 struct/class forward declaration mismatches
- C4005: Fix 4 macro redefinitions (added #ifndef guards)
- C4100: Fix 1,008 unreferenced parameters (/*paramName*/ pattern)
- C4307: Fix 1 integral constant overflow

Key patterns applied:
- Float literals with f suffix and static_cast<float>() for M_PI
- size_t for container size comparisons
- [[maybe_unused]] for intentionally unused loop variables
- Reordered member initializer lists to match declaration order
- struct/class consistency in forward declarations
- #ifndef guards for PROFILE_SCOPE, PROFILE_FUNCTION, _USE_MATH_DEFINES

Remaining 15 low-priority warnings in placeholder/stub code.
Build compiles successfully with zero errors.

+++ src/modules/Playerbot/AI/ClassAI/CombatSpecializationBase.cpp
@@ -270,7 +270,7 @@ Position CombatSpecializationBase::GetOptimalPosition(::Unit* target)
     if (_role == CombatRole::MELEE_DPS)
     {
         // Try to get behind target
-        angle = target->GetOrientation() + M_PI;
+        angle = target->GetOrientation() + static_cast<float>(M_PI);
         
     }
     else if (_role == CombatRole::TANK)
